Better status code handling and documentation
The application now handles checking for the content-type and request
method earlier in the request processing, simplifying the route method
a bit and making it a bit clearer why we're returning the code we are.

We also now return 405 and 415 for invalid requests rather than 200
with a message in the response body. If we're expecting JSON, we should
enforce that.